# src/api/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"User": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"webId": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"sessions": {
						"items": {
							"$ref": "#/components/schemas/UserSession"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"webId",
					"sessions"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserSession": {
				"properties": {
					"id": {
						"type": "string"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"session": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"user",
					"session"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Vocabulary": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"contributors": {
						"items": {
							"$ref": "#/components/schemas/User"
						},
						"type": "array"
					},
					"classes": {
						"items": {
							"$ref": "#/components/schemas/RdfClass"
						},
						"type": "array"
					},
					"name": {
						"type": "string"
					},
					"slug": {
						"type": "string"
					},
					"description": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"contributors",
					"classes",
					"name",
					"slug"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RdfClass": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"vocab": {
						"$ref": "#/components/schemas/Vocabulary"
					},
					"properties": {
						"items": {
							"$ref": "#/components/schemas/Property"
						},
						"type": "array"
					},
					"inherits": {
						"$ref": "#/components/schemas/RdfClass"
					},
					"name": {
						"type": "string"
					},
					"slug": {
						"type": "string"
					},
					"description": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"vocab",
					"properties",
					"name",
					"slug"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Property": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"domain": {
						"$ref": "#/components/schemas/RdfClass"
					},
					"name": {
						"type": "string"
					},
					"slug": {
						"type": "string"
					},
					"description": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"domain",
					"name",
					"slug"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PropertyCreationParams": {
				"properties": {
					"domain": {
						"type": "string"
					}
				},
				"required": [
					"domain"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ClassCreationParams": {
				"properties": {
					"inherits": {
						"type": "string"
					}
				},
				"required": [
					"inherits"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SolidProfileShapeType.SchemPerson": {
				"enum": [
					"http://schema.org/Person"
				],
				"type": "string"
			},
			"SolidProfileShapeType.FoafPerson": {
				"enum": [
					"http://xmlns.com/foaf/0.1/Person"
				],
				"type": "string"
			},
			"SolidProfileShape": {
				"allOf": [
					{
						"properties": {
							"knows": {
								"anyOf": [
									{
										"type": "string"
									},
									{
										"items": {
											"type": "string"
										},
										"type": "array"
									}
								]
							},
							"publicTypeIndex": {
								"anyOf": [
									{
										"type": "string"
									},
									{
										"items": {
											"type": "string"
										},
										"type": "array"
									}
								]
							},
							"privateTypeIndex": {
								"anyOf": [
									{
										"type": "string"
									},
									{
										"items": {
											"type": "string"
										},
										"type": "array"
									}
								]
							},
							"account": {
								"type": "string"
							},
							"storage": {
								"anyOf": [
									{
										"type": "string"
									},
									{
										"items": {
											"type": "string"
										},
										"type": "array"
									}
								]
							},
							"inbox": {
								"type": "string"
							},
							"img": {
								"type": "string"
							},
							"hasPhoto": {
								"type": "string"
							},
							"name": {
								"type": "string"
							},
							"fn": {
								"type": "string"
							},
							"id": {
								"type": "string"
							}
						},
						"required": [
							"inbox",
							"id"
						],
						"type": "object"
					},
					{
						"properties": {
							"type": {
								"items": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/SolidProfileShapeType.SchemPerson"
										},
										{
											"$ref": "#/components/schemas/SolidProfileShapeType.FoafPerson"
										}
									]
								},
								"type": "array"
							}
						},
						"required": [
							"type"
						],
						"type": "object"
					}
				]
			}
		},
		"securitySchemes": {
			"cookie": {
				"type": "apiKey",
				"in": "cookie",
				"name": "IPS_SESSION_ID"
			}
		}
	},
	"info": {
		"title": "express-ips",
		"version": "0.1.0",
		"license": {
			"name": "AGPL-3.0"
		},
		"contact": {
			"name": "Collaboware UG"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/vocabs": {
			"get": {
				"operationId": "GetVocabs",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Vocabulary"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/vocabs/{vocab}": {
			"get": {
				"operationId": "GetVocab",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Vocabulary"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "vocab",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateVocab",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Vocabulary"
								}
							}
						}
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "vocab",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/vocabs/{vocab}/{propertyOrClass}": {
			"get": {
				"operationId": "GetPropertyOrClass",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "vocab",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "propertyOrClass",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreatePropertyOrClass",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/Property"
										},
										{
											"$ref": "#/components/schemas/RdfClass"
										}
									]
								}
							}
						}
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "vocab",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "propertyOrClass",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"anyOf": [
									{
										"$ref": "#/components/schemas/PropertyCreationParams"
									},
									{
										"$ref": "#/components/schemas/ClassCreationParams"
									}
								]
							}
						}
					}
				}
			}
		},
		"/user/profile": {
			"get": {
				"operationId": "Profile",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/SolidProfileShape"
										},
										{
											"properties": {
												"errors": {
													"items": {
														"type": "string"
													},
													"type": "array"
												}
											},
											"required": [
												"errors"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": []
			}
		},
		"/user/logout": {
			"get": {
				"operationId": "Logout",
				"responses": {
					"200": {
						"description": "Logout successful",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/user/login": {
			"get": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Login successful"
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/user/redirect-from-solid-idp": {
			"get": {
				"operationId": "RedirectFromSolidIdp",
				"responses": {
					"301": {
						"description": "Successfully redirected from login"
					},
					"500": {
						"description": "Login was not successful."
					}
				},
				"security": [],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "http://localhost:3000/api"
		}
	]
}